requirements.txt
aiodns==3.0.0
cffi==1.14.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.0.0
pycparser==2.20
python-dotenv==0.19.0
slixmpp==1.7.1
